DEV: Implemented custom CSVWriter that adds weights to the csv output. Now it inherits from the base csv writer in force-bdss core
