--- itwm_example/csv_writer/csv_writer_model.py
@@ -1,6 +1,0 @@
-from traits.api import Unicode
-from force_bdss.api import BaseNotificationListenerModel
-
-
-class CSVWriterModel(BaseNotificationListenerModel):
-    path = Unicode("output.csv")
